refactor: move BookkeepingPeriodStatus enum into schemas
