+++ src/components/Tasks/types.ts
@@ -1,5 +1,5 @@
+import { type BookkeepingPeriodStatus } from '@schemas/bookkeepingPeriods'
 import type { UserVisibleTask } from '@utils/bookkeeping/tasks/bookkeepingTasksFilters'
-import { type BookkeepingPeriodStatus } from '@hooks/api/businesses/[business-id]/bookkeeping/periods/useBookkeepingPeriods'
 
 export type MonthData = {
   monthStr: string
