+++ src/schemas/bookkeepingPeriods.ts
@@ -1,0 +1,10 @@
+export enum BookkeepingPeriodStatus {
+  BOOKKEEPING_NOT_ACTIVE = 'BOOKKEEPING_NOT_ACTIVE',
+  NOT_STARTED = 'NOT_STARTED',
+  IN_PROGRESS_AWAITING_BOOKKEEPER = 'IN_PROGRESS_AWAITING_BOOKKEEPER',
+  IN_PROGRESS_AWAITING_CUSTOMER = 'IN_PROGRESS_AWAITING_CUSTOMER',
+  CLOSING_IN_REVIEW = 'CLOSING_IN_REVIEW',
+  CLOSING_OPEN_ITEMS = 'CLOSING_OPEN_ITEMS',
+  CLOSED_OPEN_TASKS = 'CLOSED_OPEN_TASKS',
+  CLOSED_COMPLETE = 'CLOSED_COMPLETE',
+}
